backend/requirements.txt
# FastAPI and Web Framework
fastapi
uvicorn
python-multipart

# Database
firebase-admin

# Authentication & Security
python-jose
passlib
argon2-cffi
python-dotenv

# Data Validation
pydantic
pydantic-settings
email-validator

# Analytics & Data Processing
pandas
numpy

# Machine Learning
scikit-learn
joblib

# HTTP Client
httpx

# AI/ML Services
google-generativeai
